e2e: add back verifyImgArch()
